Restructure gmail listener
