+++ .github/workflows/release.yml
@@ -28,14 +28,14 @@ jobs:
                   args:
                       build -c
               env:
-                  WORKING_DIR: ./gmail-ballerina        
+                  WORKING_DIR: ./gmail        
             - name: Ballerina Push
               uses: ballerina-platform/ballerina-action/@master
               with:
                   args:
                       push 
               env:
-                  WORKING_DIR: ./gmail-ballerina
+                  WORKING_DIR: ./gmail
                   JAVA_HOME: /usr/lib/jvm/default-jvm
                   JAVA_OPTS: -DBALLERINA_DEV_COMPILE_BALLERINA_ORG=true
                   BALLERINA_CENTRAL_ACCESS_TOKEN: ${{ secrets.BALLERINA_CENTRAL_ACCESS_TOKEN }}
