+++ build.gradle
@@ -1,8 +1,19 @@
 /*
- * This file was generated by the Gradle 'init' task.
+ * Copyright (c) 2021 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
- * This is a general purpose Gradle build.
- * Learn more about Gradle by exploring our samples at https://docs.gradle.org/7.0/samples
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 plugins {
     id "com.github.spotbugs" version "4.0.5"
@@ -53,5 +64,5 @@ subprojects {
 
 
 task build {
-    dependsOn('gmail-ballerina:build')
+    dependsOn('gmail:build')
 }
