--- gmail-native/build.gradle
@@ -1,27 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This is a general purpose Gradle build.
- * Learn more about Gradle by exploring our samples at https://docs.gradle.org/6.7.1/samples
- */
-plugins {
-    id 'java'
-}
-
-description = 'Ballerina - Gmail Java Utils'
-
-dependencies {
-    compile group: 'org.ballerinalang', name: 'ballerina-lang', version: "${ballerinaLangVersion}"
-    compile (group: 'org.ballerinalang', name: 'ballerina-runtime', version: "${ballerinaLangVersion}") {
-        transitive = false
-    }
-}
-
-compileJava {
-    doFirst {
-        options.compilerArgs = [
-                '--module-path', classpath.asPath,
-        ]
-        classpath = files()
-    }
-}
