--- gmail-native/src/main/java/module-info.java
@@ -1,4 +1,0 @@
-module io.ballerinax.googleapis.gmail {
-    requires io.ballerina.lang;
-    requires io.ballerina.runtime;
-}
