+++ gmail/Dependencies.toml
@@ -30,6 +30,11 @@ version = "1.1.0-alpha7"
 
 [[dependency]]
 org = "ballerina"
+name = "uuid"
+version = "0.10.0-alpha7"
+
+[[dependency]]
+org = "ballerina"
 name = "time"
 version = "2.0.0-alpha8"
