+++ gmail/tests/resources/test.txt
