+++ java-wrapper/src/main/java/org/ballerinalang/googleapis/gmail/HttpNativeOperationHandler.java
@@ -1,3 +1,20 @@
+/*
+ * Copyright (c) 2021 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package org.ballerinalang.googleapis.gmail;
 
 import io.ballerina.runtime.api.Environment;
@@ -20,9 +37,6 @@ import java.util.ArrayList;
 import static io.ballerina.runtime.api.utils.StringUtils.fromString;
 
 public class HttpNativeOperationHandler {
-    public static Object callOnMailboxChanges(Environment env, BObject bHttpService, BMap<BString, Object> message) {
-        return invokeRemoteFunction(env, bHttpService, message, "callOnMailboxChanges", "onMailboxChanges");
-    }
 
     public static Object callOnNewEmail(Environment env, BObject bHttpService, BMap<BString, Object> message) {
         return invokeRemoteFunction(env, bHttpService, message, "callOnNewEmail", "onNewEmail");
@@ -36,8 +50,8 @@ public class HttpNativeOperationHandler {
         return invokeRemoteFunction(env, bHttpService, message, "callOnNewLabeledEmail", "onNewLabeledEmail");
     }
 
-    public static Object callOnNewStaredEmail(Environment env, BObject bHttpService, BMap<BString, Object> message) {
-        return invokeRemoteFunction(env, bHttpService, message, "callOnNewStaredEmail", "onNewStaredEmail");
+    public static Object callOnNewStarredEmail(Environment env, BObject bHttpService, BMap<BString, Object> message) {
+        return invokeRemoteFunction(env, bHttpService, message, "callOnNewStarredEmail", "onNewStarredEmail");
     }
 
     public static Object callOnLabelRemovedEmail(Environment env, BObject bHttpService, BMap<BString, Object> message) {
