+++ java-wrapper/src/main/java/org/ballerinalang/googleapis/gmail/ModuleUtils.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2021 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.ballerinalang.googleapis.gmail;
+
+import io.ballerina.runtime.api.Environment;
+import io.ballerina.runtime.api.Module;
+
+/**
+ * {@code ModuleUtils} contains the utility methods for the module.
+ * 
+ * @since 2.0.0
+ */
+public class ModuleUtils {
+
+    private static Module httpListenerModule;
+
+    private ModuleUtils() {}
+
+    public static void setModule(Environment environment) {
+        httpListenerModule = environment.getCurrentModule();
+    }
+
+    public static Module getModule() {
+        return httpListenerModule;
+    }
+}
